Groups RecycleView + Chat Classes Added

+++ app/src/main/java/com/example/gamepartners/data/model/UserProfileFragment.java
@@ -173,7 +173,7 @@ public class UserProfileFragment extends Fragment {
                 }).addOnProgressListener(new OnProgressListener<UploadTask.TaskSnapshot>() {
             @Override
             public void onProgress(@NonNull UploadTask.TaskSnapshot snapshot) {
-                double progressPercent = (100.00 * snapshot.getBytesTransferred()/ snapshot.getTotalByteCount());
+                int progressPercent = (int)(100.0 * snapshot.getBytesTransferred()/ snapshot.getTotalByteCount());
                 uploadProgress.setMessage(progressPercent + "%");
             }
         });
